+++ src/hierarchy.cpp
@@ -7,25 +7,25 @@ using namespace Rcpp;
 //  make hierarchy from clustering
 // [[Rcpp::export]]
 List clusterInx(IntegerMatrix x) {
-    size_t n = x.ncol();
+    int n = x.ncol();
     std::vector<bool> chacked(n);
     std::map< int, std::vector<int> > clustMemb;
     std::map< int, std::vector<int> > clustLevel;
     std::map< int, int > clustNo;
     std::vector<int> inx;
-    size_t nInx = 0;
-    size_t a, b;
+    int nInx = 0;
+    int a, b;
     bool newClust = false;
-    size_t clustCount = 1;
-    size_t clustNow = 1;
+    int clustCount = 1;
+    int clustNow = 1;
     // run thru all rows
-    for (size_t k = 0; k < x.nrow(); ++ k) {
+    for (int k = 0; k < x.nrow(); ++ k) {
         // set chack to zero
-        for (size_t i = 0; i < chacked.size (); ++ i) {
+        for (int i = 0; i < chacked.size (); ++ i) {
             chacked[i] = true;
         }
         // compare element 'i' with all following elements
-        for (size_t i = 0; i < n; ++ i) {
+        for (int i = 0; i < n; ++ i) {
             if (IntegerVector::is_na(x(k, i))) { 
               // if NA nexed
               continue;
@@ -35,7 +35,7 @@ List clusterInx(IntegerMatrix x) {
                 if (chacked[i]) {
                     inx.push_back(i);
                     // if further elements belong to the same cluster add them to
-                    for (size_t j = i +1; j < n; ++ j) {
+                    for (int j = i +1; j < n; ++ j) {
                         if (x(k, i) == x(k, j)) {
                             inx.push_back(j);
                         }
@@ -45,10 +45,10 @@ List clusterInx(IntegerMatrix x) {
                     // if 'k' is in row 2 or higher, test if similar cluster 
                     // existed at previous level as well
                     if (k > 0) {
-                        for (size_t j = 0; j < n; ++ j) {
+                        for (int j = 0; j < n; ++ j) {
                             if (x(k -1, a) == x(k -1, j)) {
                                 newClust = true;
-                                for (size_t l = 0; l < nInx; ++ l) {
+                                for (int l = 0; l < nInx; ++ l) {
                                     if (j == inx[l]) {
                                         newClust = false;
                                         break;
@@ -64,7 +64,7 @@ List clusterInx(IntegerMatrix x) {
                     }
                     // if cluster is new at it to the index tree
                     if (newClust) {
-                        for (size_t j = 0; j < inx.size (); ++ j) {
+                        for (int j = 0; j < inx.size (); ++ j) {
                             b = inx[j];
                             clustMemb[clustCount].push_back(b +1); // make R index 
                             chacked[b] = false;
@@ -73,7 +73,7 @@ List clusterInx(IntegerMatrix x) {
                         clustNo[a] = clustCount;
                         ++ clustCount; 
                     } else {
-                        for (size_t j = 0; j < inx.size (); ++ j) {
+                        for (int j = 0; j < inx.size (); ++ j) {
                             b = inx[j];
                             chacked[b] = false;
                         }
