+++ src/hit.cpp
@@ -7,7 +7,7 @@ using namespace Rcpp;
 // find significant cluster from p-value matrix
 // [[Rcpp::export]]
 NumericMatrix sigCluster(NumericVector pVal, List clustLevel, List clustMemb,
-                         const size_t p, const double alpha, 
+                         const int p, const double alpha, 
                          int minDistInx, int maxDistInx) 
 {
     // R Indices to C++
@@ -15,13 +15,13 @@ NumericMatrix sigCluster(NumericVector pVal, List clustLevel, List clustMemb,
     maxDistInx --;
     // Initialisation
     NumericMatrix sigClust(2, p);
-    for (size_t l = 0; l < p; ++ l) {
+    for (int l = 0; l < p; ++ l) {
         sigClust(1, l) = 1;
     }
     IntegerVector gIDs;
     std::vector<int> gInx;
-    size_t k = 1;
-    size_t m, h;
+    int k = 1;
+    int m, h;
     bool newClust = true;
     bool isZero = false;
     bool isNotZero = false;
@@ -29,12 +29,12 @@ NumericMatrix sigCluster(NumericVector pVal, List clustLevel, List clustMemb,
     for (int i = minDistInx; i > maxDistInx; -- i) {
         gIDs = clustLevel[i];
         // test each cluster at dist i
-        for (size_t j = 0; j < gIDs.size(); ++ j) {
+        for (int j = 0; j < gIDs.size(); ++ j) {
             m = gIDs[j] -1; // R Indix to C++
             gInx = clustMemb[m]; 
             // test if all variables of cluster have already been significant 
             // before at smaller distances
-            for (size_t l = 0; l < gInx.size(); ++ l) {
+            for (int l = 0; l < gInx.size(); ++ l) {
                 h = gInx[l] -1; // R Indix to C++
                 if (sigClust(0, h) == 0) {
                     isZero = true;
@@ -51,7 +51,7 @@ NumericMatrix sigCluster(NumericVector pVal, List clustLevel, List clustMemb,
             }
             // if newClust == true make sig cluster
             if (newClust) {
-                for(size_t l = 0; l < gInx.size(); ++ l) {
+                for(int l = 0; l < gInx.size(); ++ l) {
                     h = gInx[l] -1; // R Indix to C++
                     sigClust(0, h) = k;
                     sigClust(1, h) = pVal[m];
@@ -66,8 +66,8 @@ NumericMatrix sigCluster(NumericVector pVal, List clustLevel, List clustMemb,
     }
     // ordert cluster index 
     NumericVector cInx = sort_unique(sigClust(0, _));
-    for (size_t i = 0; i < cInx.size(); ++ i) {
-        for (size_t l = 0; l < p; ++ l) {
+    for (int i = 0; i < cInx.size(); ++ i) {
+        for (int l = 0; l < p; ++ l) {
             if (sigClust(0, l) == cInx[i]) {
                 sigClust(0, l) = i;
             }
